requirements (3).txt
python-telegram-bot[job-queue]==21.6
requests==2.32.3
httpx==0.27.2
